Avoid extra static_casts

+++ src/fheroes2/ai/ai_hero_action.cpp
@@ -1418,7 +1418,7 @@ namespace
 
         hero.ResetMovePoints();
         hero.Move2Dest( dst_index );
-        from.setBoat( Maps::GetDirection( fromIndex, dst_index ), static_cast<uint8_t>( hero.GetColor() ) );
+        from.setBoat( Maps::GetDirection( fromIndex, dst_index ), hero.GetColor() );
         hero.SetShipMaster( false );
         hero.GetPath().Reset();
